Moved OCSPResponderFixture to a test lib
